requirements.txt
discord.py>=2.4.0
ccxt>=4.3.0
pandas>=2.2.3
numpy>=1.26.0
python-dotenv>=1.0.1
pyyaml>=6.0.2
matplotlib
pillow
# Solo si discord.py se queja de audioop en Python 3.13:
audioop-lts; python_version >= "3.13"
